+++ src/main/java/com/travellab/repository/TicketReturnRepository.java
@@ -6,5 +6,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.rest.core.annotation.RepositoryRestResource;
 
 @RepositoryRestResource()
-public interface TicketReturnRepository extends JpaRepository<TicketReturn, Integer> {
+public interface TicketReturnRepository extends JpaRepository<TicketReturn, Long> {
 }
